feat(backend): Create toy database
The sole focus is on developing the core features of the API. Part of
this will require a toy database to play around with before I create the
SQLite database that I plan to build.

The `compose.yaml` file has been limited to just the backend to cut down
on unnecessary build times.